Improve share extension init
