+++ app/share.js
@@ -6,14 +6,10 @@ import RNUserDefaults from 'rn-user-defaults';
 
 import Navigation from './lib/Navigation';
 import store from './lib/createStore';
-import database from './lib/realm';
 import sharedStyles from './views/Styles';
 import { isNotch, isIOS } from './utils/deviceInfo';
 import { defaultHeader, onNavigationStateChange } from './utils/navigation';
-import { selectServerSuccess } from './actions/server';
-import { setUser } from './actions/login';
 import RocketChat from './lib/rocketchat';
-import { addSettings } from './actions';
 
 const InsideNavigator = createStackNavigator({
 	ShareListView: {
@@ -66,25 +62,7 @@ class Root extends React.Component {
 		const currentServer = await RNUserDefaults.get('currentServer');
 
 		if (currentServer) {
-			// server
-			database.setActiveDB(currentServer);
-			const { serversDB } = database.databases;
-			const server = serversDB.objectForPrimaryKey('servers', currentServer);
-			store.dispatch(selectServerSuccess(currentServer, server.version));
-
-			// add settings to upload media
-			store.dispatch(addSettings({
-				Site_Url: server.id,
-				useRealName: server.useRealName,
-				FileUpload_MediaTypeWhiteList: server.FileUpload_MediaTypeWhiteList,
-				FileUpload_MaxFileSize: server.FileUpload_MaxFileSize
-			}));
-
-			// user
-			const userId = await RNUserDefaults.get(`${ RocketChat.TOKEN_KEY }-${ currentServer }`);
-			const user = userId && serversDB.objectForPrimaryKey('user', userId);
-			store.dispatch(setUser(user));
-
+			await RocketChat.shareExtensionInit(currentServer);
 			await Navigation.navigate('InsideStack');
 		} else {
 			await Navigation.navigate('OutsideStack');
