+++ app/views/SelectServerView.js
@@ -5,17 +5,11 @@ import {
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { SafeAreaView } from 'react-navigation';
-import RNUserDefaults from 'rn-user-defaults';
 
 import I18n from '../i18n';
 import database from '../lib/realm';
 import StatusBar from '../containers/StatusBar';
-import { selectServerSuccess as selectServerSuccessAction } from '../actions/server';
-import { setUser as setUserAction } from '../actions/login';
-import { addSettings as addSettingsAction } from '../actions';
-import {
-	COLOR_BACKGROUND_CONTAINER
-} from '../constants/colors';
+import { COLOR_BACKGROUND_CONTAINER } from '../constants/colors';
 import Navigation from '../lib/Navigation';
 import ServerItem, { ROW_HEIGHT } from '../presentation/ServerItem';
 import sharedStyles from './Styles';
@@ -41,10 +35,6 @@ const styles = StyleSheet.create({
 
 @connect(state => ({
 	server: state.server.server
-}), dispatch => ({
-	setUser: user => dispatch(setUserAction(user)),
-	addSettings: settings => dispatch(addSettingsAction(settings)),
-	selectServerSuccess: server => dispatch(selectServerSuccessAction(server))
 }))
 export default class SelectServerView extends React.Component {
 	static navigationOptions = () => ({
@@ -52,10 +42,7 @@ export default class SelectServerView extends React.Component {
 	})
 
 	static propTypes = {
-		server: PropTypes.string,
-		setUser: PropTypes.func,
-		addSettings: PropTypes.func,
-		selectServerSuccess: PropTypes.func
+		server: PropTypes.string
 	}
 
 	constructor(props) {
@@ -70,29 +57,11 @@ export default class SelectServerView extends React.Component {
 
 	select = async(server) => {
 		const {
-			server: currentServer, selectServerSuccess, setUser, addSettings
+			server: currentServer
 		} = this.props;
 
 		if (currentServer !== server) {
-			const { serversDB } = database.databases;
-
-			// server
-			database.setActiveDB(server);
-			const serverInfo = serversDB.objectForPrimaryKey('servers', server);
-			selectServerSuccess(server, serverInfo.version);
-
-			// add settings to upload media
-			addSettings({
-				Site_Url: serverInfo.id,
-				useRealName: serverInfo.useRealName,
-				FileUpload_MediaTypeWhiteList: serverInfo.FileUpload_MediaTypeWhiteList,
-				FileUpload_MaxFileSize: serverInfo.FileUpload_MaxFileSize
-			});
-
-			// user
-			const userId = await RNUserDefaults.get(`${ RocketChat.TOKEN_KEY }-${ server }`);
-			const user = userId && serversDB.objectForPrimaryKey('user', userId);
-			setUser(user);
+			await RocketChat.shareExtensionInit(server);
 		}
 		Navigation.navigate('ShareListView');
 	}
